+++ src/handlers/parsers.js
@@ -597,6 +597,12 @@ export const parseNewTransaction = async (
   nativeCurrency = '',
 ) => {
   let _transactions = [...transactions];
+  let ethFeeAsset = {
+    name: 'Ethereum',
+    symbol: 'ETH',
+    address: null,
+    decimals: 18,
+  };
 
   let totalGas = multiply(txDetails.gasLimit, txDetails.gasPrice);
   let txFee = {
@@ -635,7 +641,11 @@ export const parseNewTransaction = async (
   const assetSymbol = tx.asset.symbol;
   tx.native = { selected: nativeCurrencies[nativeCurrency] };
 
-  const response = await apiGetHistoricalPrices(assetSymbol, timestamp);
+  const priceAssets = [assetSymbol, 'ETH'];
+  const promises = priceAssets.map(x => apiGetHistoricalPrices(x, timestamp));
+  const historicalPriceResponses = await Promise.all(promises);
+  const response = historicalPriceResponses[0];
+  const feeResponse = historicalPriceResponses[1];
 
   if (response.data.response !== 'Error' && response.data[assetSymbol]) {
     await Promise.all(
@@ -643,16 +653,24 @@ export const parseNewTransaction = async (
         const assetPriceAmount = convertAmountToBigNumber(
           response.data[assetSymbol][nativeCurrency],
         );
+        const feePriceAmount = convertAmountToBigNumber(
+          feeResponse.data['ETH'][nativeCurrency],
+        );
         let prices = { selected: nativeCurrencies[nativeCurrency] };
         prices[nativeCurrency] = {};
         prices[nativeCurrency][assetSymbol] = {
           price: { amount: assetPriceAmount, display: null },
         };
+        prices[nativeCurrency]['ETH'] = {
+          price: { amount: feePriceAmount, display: null },
+        };
         const assetPriceDisplay = convertAmountToDisplay(
           assetPriceAmount,
           prices,
         );
+        const feePriceDisplay = convertAmountToDisplay(feePriceAmount, prices);
         prices[nativeCurrency][assetSymbol].price.display = assetPriceDisplay;
+        prices[nativeCurrency]['ETH'].price.display = feePriceDisplay;
         const assetPrice = prices[nativeCurrency][assetSymbol].price;
         const valuePriceAmount = convertAssetAmountToNativeValue(
           tx.value.amount,
@@ -669,7 +687,7 @@ export const parseNewTransaction = async (
           : { amount: '', display: '' };
         const txFeePriceAmount = convertAssetAmountToNativeValue(
           tx.txFee.amount,
-          tx.asset,
+          ethFeeAsset,
           prices,
         );
         const txFeePriceDisplay = convertAmountToDisplay(
@@ -879,6 +897,12 @@ export const parseAccountTransactions = async (
  */
 export const parseHistoricalPrices = async (transactions = null) => {
   if (!transactions.length) return transactions;
+  const ethFeeAsset = {
+    name: 'Ethereum',
+    symbol: 'ETH',
+    address: null,
+    decimals: 18,
+  };
   const _transactions = await Promise.all(
     transactions.map(async (tx, idx) => {
       const timestamp = tx.timestamp ? tx.timestamp.secs : Date.now();
@@ -888,15 +912,19 @@ export const parseHistoricalPrices = async (transactions = null) => {
       }
       if (!tx.native || (tx.native && Object.keys(tx.native).length < 1)) {
         try {
-          const response = await debounceRequest(
-            apiGetHistoricalPrices,
-            [assetSymbol, timestamp],
-            100 * idx,
+          const priceAssets = [assetSymbol, 'ETH'];
+          const promises = priceAssets.map(x =>
+            debounceRequest(apiGetHistoricalPrices, [x, timestamp], 100 * idx),
           );
+          const historicalPriceResponses = await Promise.all(promises);
+          const response = historicalPriceResponses[0];
+          const feeResponse = historicalPriceResponses[1];
 
           if (
             response.data.response === 'Error' ||
-            !response.data[assetSymbol]
+            !response.data[assetSymbol] ||
+            feeResponse.data.response === 'Error' ||
+            !feeResponse.data['ETH']
           ) {
             return tx;
           }
@@ -917,6 +945,18 @@ export const parseHistoricalPrices = async (transactions = null) => {
             prices[nativeCurrency][
               assetSymbol
             ].price.display = assetPriceDisplay;
+
+            const feePriceAmount = convertAmountToBigNumber(
+              feeResponse.data['ETH'][nativeCurrency],
+            );
+            prices[nativeCurrency]['ETH'] = {
+              price: { amount: feePriceAmount, display: null },
+            };
+            const feePriceDisplay = convertAmountToDisplay(
+              feePriceAmount,
+              prices,
+            );
+            prices[nativeCurrency]['ETH'].price.display = feePriceDisplay;
             const assetPrice = prices[nativeCurrency][assetSymbol].price;
             let asset = { ...tx.asset };
             if (asset.symbol === 'WETH') {
@@ -936,7 +976,7 @@ export const parseHistoricalPrices = async (transactions = null) => {
               : null;
             const txFeePriceAmount = convertAssetAmountToNativeValue(
               tx.txFee.amount,
-              asset,
+              ethFeeAsset,
               prices,
               tx,
             );