+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added
 
+* Versioning for local storage
 * Bookmark Reminder (anti-phishing) Ribbon
 * Exchange button and modal
 * Shapeshift endpoints
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Changed
 
+* Fix txn fee conversion
 * Layout padding definition locations
 * Fix mobile layout on the wallet selection screen (#232)
 * DrodpownAsset improved logic